Incluida funcionalidade que importa dados CSV e salva em um banco de dados

+++ poligonos/src/main/java/br/com/metrificacao/poligonos/dto/PontoPoligonoDto.java
@@ -9,9 +9,9 @@ public class PontoPoligonoDto {
 
     private Long id;
 
-    private Long coordenadaX;
+    private Double coordenadaX;
 
-    private Long coordenadaY;
+    private Double coordenadaY;
 
     private String nomePoligono;
